Add /rancher, Abstract cmds.py, Add DirectResponse model
Well jazzmin is down anyway, but if she were up, she would have been able to send me a GET request to let me know that an error occurred.

Added a DirectReponse class to models.py so that the /rancher endpoint would send a text to my phone without being prompted to from iMessage. I imagine this would work for other stuff too.

Added a do_command() method in cmds.py so that we aren't re-using code to create a cmdline process.