Fix !wolf response when res.details DNE
I tried out !wolf anime, only to find that it yields an abnormal result, where the .details property does not exist.
To fix this, we look to see if there are any non-@-prefixed keys (and also not 'attributes'), and then add those keys and their pretty-formatted values to our response list.